+++ packages/_private/vue-demo/src/HomeView.vue
@@ -1,8 +1,10 @@
 <template>
-<h1>
-  Welcome to the Synergy Vue Demo!
-</h1>
+  <h1 class="syn-heading--3x-large">
+    Welcome to the Synergy Vue Demo!
+  </h1>
 
-In this demo some use cases and our components are tested in the Vue ecosystem, like navigation handling, forms and theme switching.
-
+  <p class="syn-body--medium">
+    In this demo some use cases and our components are tested in the Vue ecosystem,
+    like navigation handling, forms and theme switching.
+  </p>
 </template>