+++ packages/components/scripts/jobs/cem.js
@@ -1,11 +1,16 @@
-import { execPromise, formatFile, getPath, job } from './shared.js';
+import {
+  execPromise,
+  formatFile,
+  getPath,
+  job,
+} from './shared.js';
 
 /**
  * Run cem for the components package
  */
-export const runCem = job('Creating component manifest...', async () => {
+export const runCem = job('Creating component manifest...', async (componentDistDir = 'dist') => {
   await execPromise(
-    'cem analyze --litelement --outdir dist',
+    `cem analyze --litelement --outdir ${componentDistDir}`,
     { stdio: 'inherit' },
   );