+++ packages/components/scripts/jobs/index.js
@@ -6,5 +6,6 @@ export * from './createExports.js';
 export * from './esbuildComponents.js';
 export * from './prepare.js';
 export * from './react.js';
+export * from './styles.js';
 export * from './tsc.js';
 export * from './vue.js';